Get tibble of feedbacks.

Returns a tibble of feedback which can be a straight list or grouped by contact using the group_by parameter. The list can also be filtered by state query parameter for active, archived, unsorted or all.Note that by default only active feedback is returned.

Usage

pp_get_feedbacks(
  state = c("all", "active", "archived", "unsorted"),
  company = NULL,
  company_country = NULL,
  company_size = NULL,
  company_value = NULL,
  customer = NULL,
  product = NULL,
  persona = NULL,
  job_role = NULL,
  tags = NULL,
  has_ideas = NULL,
  external_id = NULL,
  external_url = NULL,
  group_by = FALSE,
  page = 1,
  size = 100,
  .limit = NULL,
  ...
)

Arguments

state

Set to value of active for active, archived for archived feedback, unsorted for unsorted feedback and all for all feedback. Default is all.

company

Set to filter the feedback results based on whether the feedback was entered for a contact linked to the company.

company_country

Set to filter the feedback results based on the country set for the associated company.

company_size

Set to filter the feedback results based on the size of the company for the associated company.

company_value

Set to filter the feedback results based on the value of the company for the company associated to the feedback.

customer

Can be either numeric ID, UUID of a contact or contact email.

product

Filter results by the product associated to feedback. Can either be the product UUID or product ID.

persona

Filter results by the persona associated to feedback. Can either be the persona UUID or persona ID.

job_role

Filter results by the job role of the contact associated to the feedback. Use the JobRole UUID.

tags

Filter feedback results by the tags associated to the feedback. Multiple tags can be specified and acts as an OR. Use the tag ID or UUID.

has_ideas

Whether the feedback is associated to one or more ideas. Default is either.

external_id

Filter feedback to return the feedback associated with a specific External ID. An example of an external ID is the ID of a record in a CRM or ID of a ticket in a customer support application.

external_url

Filter feedback to return the feedback associated with a specific external url. An example of an external url is that of a record in a CRM or a ticket in a customer support application

group_by

Setting to customer will group the feedback by customer

page

Set to page through the results. Default is 1.

size

For setting the number of results per page. Default is 100.

.limit

The maximum number of elements to return

...

Other arguments passed to .pp()]

Value

tibble

Note

GET /feedbacks