Get an idea.

Use the numeric ID to return an idea. You can get a simple version of the idea or an expanded one using the expanded parameter which returns the business case, functional specs, notes, user stories, comments and other linked data.

Usage

pp_get_idea(id = NULL, expand = FALSE, by_project_id = FALSE, ...)

Arguments

id

Numeric ID of the idea unless using the by_project_id then it is the numeric project ID (the Idea number seen in the UI).

expand

Whether to return the expanded version of the idea object. This adds in business case, user stories, comments, etc.

by_project_id

Whether the ID represents the project_id instead of the numeric ID.

...

Other arguments passed to .pp()]

Details

If you have the project_id (the numerical number from the ProdPad UI in the idea canvas url) you can use that to fetch the idea by using the query parameter by_project_id, otherwise you need to use the global numerical ID.

Note

GET /ideas/{id}