Create a new feedback.

Create a new feedback in the account. The feedback has to have either a name or an email and the actual feedback. The email is used to add the feedback to an existing contact.

Usage

pp_post_feedback(
  contact_id,
  contact_name = NULL,
  product_id,
  product_name = NA,
  feedback,
  ...
)

Arguments

contact_id

string($uuid) ID of the contact providing the feedback. Either Contact ID or Contact name is required.

contact_name

string Name of the contact providing the feedback. Either Contact ID or Contact name is required.

product_id

Product id. See pp_get_products()

product_name

Product name. See pp_get_products()

feedback

string (Required) The feedback. This field accepts HTML and is stored as UTF-8.

...

Other arguments passed to .pp()]

Details

Fields to create a feedback. Note either the contact ID or contact name is required. Otherwise a 400 error will be returned. If a name and email field are used, the value in the email field will be used to see if the contact already exists and if it does the feedback will be automatically added to the existing contact.

Note

POST /feedbacks