Get a tibble of companies.

Return a list of companies. Companies are a collection of contacts that are providing feedback. Using companies allows you to group feedback from the same organisational source.

Usage

pp_get_companies(
  name = NULL,
  country = NULL,
  company_size = NULL,
  value = NULL,
  city = NULL,
  tags = NULL,
  external_id = NULL,
  external_url = NULL,
  contacts = TRUE,
  feedbacks = NULL,
  page = NULL,
  ...
)

Arguments

name

Filter the companies by the name or partial name of the companies.

country

Set to filter the companies based on the country. Use ISO Alpha-2 country codes. Only one country can be filtered at a time.

company_size

Set to filter the companies based on their size.

value

Set to filter the companies based on their value.

city

Set to filter the companies based on city.

tags

Filter companies by the tags associated to the feedback. Mulitple tags can be specified and acts as an OR. Use the tag ID or UUID.

external_id

Filter the companies by an ID from a 3rd party application associated to the companies in Prodpad

external_url

Filter the companies by a URL that is associated to a company.

contacts

Whether to include contacts associated with each company in the results. Default is true.

feedbacks

Whether to include the feedback for each contact associated to a company in the results. Default is true. Note if this is set to true then contacts will be returned whether or not contacts is set to true or false.

page

The page of results to return (size is always 100).

...

Other arguments passed to .pp()]

Details

While very useful for business to busines,s companies can be used to create a any sort grouping of contacts for example by a conference or a cohort.

Using parameters allows you to also select whether to return feedback associated to the company via the contacts connected to the company. This is only available to accounts with an Advanced+ subscription.

Note

GET /companies

See also